+++ src/components/Search/Search.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./landing.css";
+import "../../styles/landing.css";
 
 const Search = () => {
   return (
