+++ src/hooks/useAsync.js
@@ -4,8 +4,12 @@ export function useAsync(asyncFunction) {
   const [userEmail, setUserEmail] = useState(false);
 
   const wrappedFunction = async () => {
-    const { email } = await asyncFunction();
-    setUserEmail(email);
+    try {
+      const { email } = await asyncFunction();
+      setUserEmail(email);
+    } catch (error) {
+      console.error(error);
+    }
   };
 
   return [userEmail, wrappedFunction];
