+++ src/styles/global.css
